requirements.txt
aiodns==2.0.0
aiohttp==3.7.4.post0
aniso8601==9.0.1
async-timeout==3.0.1
attrs==20.3.0
ccxt==1.47.40
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
cryptography==3.4.7
Flask==1.1.2
Flask-RESTful==0.3.8
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
kucoin-python==1.0.6
MarkupSafe==1.1.1
multidict==5.1.0
numpy==1.20.2
pandas==1.2.3
pycares==3.1.1
pycparser==2.20
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
six==1.15.0
typing-extensions==3.7.4.3
urllib3==1.26.4
websockets==8.1
Werkzeug==1.0.1
yarl==1.1.0